addressing todos in the comments
