+++ app/sentinel/internal/safe/config.go
@@ -1,0 +1,38 @@
+/*
+|    Protect your secrets, protect your sensitive data.
+:    Explore VMware Secrets Manager docs at https://vsecm.com/
+</
+<>/  keep your secrets... secret
+>/
+<>/' Copyright 2023-present VMware Secrets Manager contributors.
+>/'  SPDX-License-Identifier: BSD-2-Clause
+*/
+
+package safe
+
+import (
+	data "github.com/vmware-tanzu/secrets-manager/core/entity/data/v1"
+	"github.com/vmware-tanzu/secrets-manager/core/env"
+)
+
+func decideBackingStore(backingStore string) data.BackingStore {
+	switch data.BackingStore(backingStore) {
+	case data.File:
+		return data.File
+	case data.Memory:
+		return data.Memory
+	default:
+		return env.BackingStoreForSafe()
+	}
+}
+
+func decideSecretFormat(format string) data.SecretFormat {
+	switch data.SecretFormat(format) {
+	case data.Json:
+		return data.Json
+	case data.Yaml:
+		return data.Yaml
+	default:
+		return data.Json
+	}
+}
