+++ app/sentinel/internal/safe/tls.go
@@ -1,0 +1,30 @@
+/*
+|    Protect your secrets, protect your sensitive data.
+:    Explore VMware Secrets Manager docs at https://vsecm.com/
+</
+<>/  keep your secrets... secret
+>/
+<>/' Copyright 2023-present VMware Secrets Manager contributors.
+>/'  SPDX-License-Identifier: BSD-2-Clause
+*/
+
+package safe
+
+import (
+	"github.com/pkg/errors"
+	"github.com/spiffe/go-spiffe/v2/spiffeid"
+	"github.com/spiffe/go-spiffe/v2/spiffetls/tlsconfig"
+	"github.com/vmware-tanzu/secrets-manager/core/validation"
+)
+
+func createAuthorizer() tlsconfig.Authorizer {
+	return tlsconfig.AdaptMatcher(func(id spiffeid.ID) error {
+		if validation.IsSafe(id.String()) {
+			return nil
+		}
+
+		return errors.New("Post: I don't know you, and it's crazy: '" +
+			id.String() + "'",
+		)
+	})
+}
